12-Oct-2016
12-Oct-2016

+++ src/main/java/navspringboot/models/Filter.java
@@ -2,7 +2,7 @@ package navspringboot.models;
 
 public class Filter {
     private String filterobjecttype;
-    private int filterobjectid;
+    private Long filterobjectid;
 
     public String getFilterobjecttype() {
         return filterobjecttype;
@@ -12,11 +12,11 @@ public class Filter {
         this.filterobjecttype = filterobjecttype;
     }
 
-    public int getFilterobjectid() {
+    public Long getFilterobjectid() {
         return filterobjectid;
     }
 
-    public void setFilterobjectid(int filterobjectid) {
+    public void setFilterobjectid(Long filterobjectid) {
         this.filterobjectid = filterobjectid;
     }
 }
